P1.3e: Split adapters/service.go into per-concern files (#39)
Pure restructuring; no behavior change. The 1,126-LOC service.go is split
into five files in the same package, all under 300 LOC:

- service.go (25)        AdapterService struct + NewAdapterService
- create.go (288)        CreateAdapter, hasStdioPackage, generateSecureToken
- lifecycle.go (105)     GetAdapter, ListAdapters, UpdateAdapter, DeleteAdapter
- templates.go (292)     command template processing + secret lookup helpers
- sidecar_meta.go (199)  getSidecarConfig + getSidecarMeta (+ sidecarMeta type)
- capabilities.go (258)  SyncAdapterCapabilities, CheckAdapterHealth,
                         discoverCapabilities*, sidecar health checks

Verbatim move of bodies, comments, and debug logging. All methods remain on
*AdapterService so package-private state and the existing test surface
(hasStdioPackage, getSidecarMeta, CreateAdapter) are unchanged.

Verified:
- go build ./... clean
- go test ./pkg/services/adapters/... passes
- go vet ./... shows only pre-existing baseline failures
  (pkg/mcp/message_router_test.go AuthInfo,
   internal/service/authorization_test.go FindByMCPSessionID)

Closes #8
